Remove lots of dead code
