+++ src/map-director/slopes-group-filler/FlatTilesetTile.cpp
@@ -18,7 +18,7 @@
 
 *****************************************************************************/
 
-#include "FlatTilesetTileN.hpp"
+#include "FlatTilesetTile.hpp"
 
 #include "../MapTileset.hpp"
