+++ src/map-director/slopes-group-filler/OutWallTilesetTile.cpp
@@ -1,7 +1,7 @@
 /******************************************************************************
 
     GPLv3 License
-    Copyright (c) 2023 Aria Janke
+    Copyright (c) 2024 Aria Janke
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -18,7 +18,7 @@
 
 *****************************************************************************/
 
-#include "OutWallTilesetTileN.hpp"
+#include "OutWallTilesetTile.hpp"
 
 namespace {
